Fixes the user serialization bug correctly
Switching our user_serializer to use camelCased keys was a mistake. It was
causing the typical serialization of our user models to break. What we needed to 
do (and what this commit does) is to normalize our user JSON before pushing it 
into our store. Checkout extras.js => Sis#normalizeJsonObject for more info.