Updates subtasks to show assigned students as tags
Removes 'Assigned to:' text + user names and adds the tags that we use on editing to the subtask show view. Does a bunch of cleaning up of the tags css as it was kinda all over the place (Sorry I guess?). Updates user_select_view.js to make sure clicking on the tags only adds/removes them when the user is editing. Fixes #59.