Updates Auth#logout to cause a page refresh
Does away with the whole shitstorm of trying to clear/update the current user after logout by logging the user out and then redirecting them to home page. This reloads everything, but it's worth it as it's way better than the bugs we were dealing with. Gets rid of sessions_controller#destroy as we no longer need to return the new CSRF token and what not. Fixes #30.