Updates comments to allow for both users.
Changes comment.js to belong to a user so that teachers can comment and dispute. Also fixes #90 as we were setting the comment.student instead of comment.user. Note that this is kind of a hack where we are then fetching a new users instead of using the currently loaded student or teacher. It's not too big of a deal as this is a small deoptimization, but if I was more serious about this project I wouldn't be adding it. :sweat_smile: